test(jest): move logger loading coverage to core
Keep Jest module-engine, resolver, mocking, and native-require bypass coverage in the core integration suite rather than the Test Management suite. Move the logger sandbox dependencies with the tests.
