+++ integration-tests/jest/jest.test-management.spec.js
@@ -61,9 +61,9 @@ const requestedJestVersion = process.env.JEST_VERSION || 'latest'
 const oldestJestVersion = DD_MAJOR >= 6 ? '28.0.0' : '24.8.0'
 const JEST_VERSION = requestedJestVersion === 'oldest' ? oldestJestVersion : requestedJestVersion
 const onlyLatestIt = JEST_VERSION === 'latest' ? it : it.skip
-const esmIt = JEST_VERSION === 'latest' || Number(JEST_VERSION.split('.')[0]) >= 28 ? it : it.skip
 const shouldInstallJestEnvironmentJsdom = JEST_VERSION === 'latest' || Number(JEST_VERSION.split('.')[0]) >= 28
 
+// TODO: add ESM tests
 describe(`jest@${JEST_VERSION} commonJS`, () => {
   let receiver
   let childProcess
@@ -78,11 +78,8 @@ describe(`jest@${JEST_VERSION} commonJS`, () => {
     JEST_VERSION !== 'latest' ? `jest-circus@${JEST_VERSION}` : '',
     ...getBabelDependencies(JEST_VERSION),
     '@happy-dom/jest-environment',
-    'bunyan',
     'jest-image-snapshot',
     'office-addin-mock',
-    'pino',
-    'winston',
   ].filter(Boolean), true)
 
   before(function () {
@@ -3556,291 +3553,6 @@ describe(`jest@${JEST_VERSION} commonJS`, () => {
         assert.strictEqual(exitCode[0], 0)
       })
     })
-  })
-
-  context('winston mocking', () => {
-    it('should allow winston to be mocked and verify createLogger is called', async () => {
-      childProcess = exec(
-        runTestsCommand,
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            TESTS_TO_RUN: 'jest-mock-bypass-require/winston-mock-test',
-            SHOULD_CHECK_RESULTS: '1',
-          },
-        }
-      )
-
-      const [code] = await once(childProcess, 'exit')
-      assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}`)
-    })
-  })
-
-  context('Pino and Bunyan module loading', () => {
-    for (const loggerName of ['pino', 'bunyan']) {
-      for (const resolutionType of ['moduleNameMapper', 'custom resolver']) {
-        it(`respects Jest ${resolutionType} for ${loggerName}`, async () => {
-          let testOutput = ''
-          // Ensure the native bypass still defers to Jest when its resolution is customized.
-          const resolutionConfig = resolutionType === 'moduleNameMapper'
-            ? {
-                CONFIG_MODULE_NAME_MAPPER: JSON.stringify({
-                  [`^${loggerName}$`]: '<rootDir>/ci-visibility/jest-mock-bypass-require/mapped-logger.js',
-                }),
-              }
-            : {
-                CONFIG_RESOLVER: '<rootDir>/ci-visibility/jest-mock-bypass-require/logger-resolver.js',
-              }
-
-          childProcess = exec(
-            runTestsCommand,
-            {
-              cwd,
-              env: {
-                ...getCiVisAgentlessConfig(receiver.port),
-                ...resolutionConfig,
-                TEST_LOGGER: loggerName,
-                TESTS_TO_RUN: 'jest-mock-bypass-require/mapped-logger-test',
-                USE_CONFIG_FILE: '1',
-                USE_JEST_RUN: '1',
-              },
-            }
-          )
-          childProcess.stdout.on('data', chunk => {
-            testOutput += chunk.toString()
-          })
-          childProcess.stderr.on('data', chunk => {
-            testOutput += chunk.toString()
-          })
-
-          const [code] = await once(childProcess, 'exit')
-          assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-        })
-      }
-
-      it(`instruments ${loggerName} after another suite mocks it`, async () => {
-        let testOutput = ''
-        const logsPromise = receiver
-          .gatherPayloadsMaxTimeout(({ url }) => url.includes('/api/v2/logs'), payloads => {
-            assert.strictEqual(payloads.length, 1, testOutput)
-
-            const [{ headers, logMessage, url }] = payloads
-            assert.strictEqual(headers['content-type'], 'application/json')
-            assert.strictEqual(headers['dd-api-key'], 'api-key')
-            assert.strictEqual(url, `/api/v2/logs?ddsource=${loggerName}&service=my-service`)
-            assert.strictEqual(logMessage.length, 1)
-
-            const [{ dd, msg }] = logMessage
-            assert.strictEqual(msg, 'real logger after mock')
-            assert.strictEqual(dd.service, 'my-service')
-            assert.match(dd.trace_id, /^\d+$/)
-            assert.match(dd.span_id, /^\d+$/)
-          })
-
-        childProcess = exec(
-          runTestsCommand,
-          {
-            cwd,
-            env: {
-              ...getCiVisAgentlessConfig(receiver.port),
-              DD_AGENTLESS_LOG_SUBMISSION_ENABLED: '1',
-              DD_AGENTLESS_LOG_SUBMISSION_URL: `http://localhost:${receiver.port}`,
-              DD_API_KEY: 'api-key',
-              DD_SERVICE: 'my-service',
-              TEST_LOGGER: loggerName,
-              // Run the mocked suite first to expose bypass state leaking between Jest runtimes.
-              TEST_SEQUENCER: './ci-visibility/jest-mock-bypass-require/test-sequencer.js',
-              TESTS_TO_RUN: `jest-mock-bypass-require/(${loggerName}-mock|z-real-logger)-test`,
-              USE_JEST_RUN: '1',
-            },
-          }
-        )
-        childProcess.stdout.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-        childProcess.stderr.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-
-        const [[code]] = await Promise.all([
-          once(childProcess, 'exit'),
-          logsPromise,
-        ])
-
-        assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-      })
-    }
-  })
-
-  context('ESM logger loading', () => {
-    for (const resolutionType of ['moduleNameMapper', 'custom resolver']) {
-      esmIt(`respects Jest ESM ${resolutionType}`, async () => {
-        let testOutput = ''
-        const resolutionConfig = resolutionType === 'moduleNameMapper'
-          ? {
-              CONFIG_MODULE_NAME_MAPPER: JSON.stringify({
-                '^winston$': '<rootDir>/ci-visibility/jest-mock-bypass-require/mapped-logger.js',
-              }),
-            }
-          : {
-              CONFIG_RESOLVER: '<rootDir>/ci-visibility/jest-mock-bypass-require/logger-resolver.js',
-            }
-
-        childProcess = exec(
-          runTestsCommand,
-          {
-            cwd,
-            env: {
-              ...getCiVisAgentlessConfig(receiver.port),
-              ...resolutionConfig,
-              CONFIG_TEST_MATCH: '**/ci-visibility/jest-mock-bypass-require/esm-mapped-logger-test.mjs',
-              NODE_OPTIONS: '-r dd-trace/ci/init --experimental-vm-modules',
-              TEST_LOGGER: 'winston',
-              USE_CONFIG_FILE: '1',
-              USE_JEST_RUN: '1',
-            },
-          }
-        )
-        childProcess.stdout.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-        childProcess.stderr.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-
-        const [code] = await once(childProcess, 'exit')
-        assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-      })
-    }
-
-    esmIt('does not resolve loggers for unrelated CommonJS ESM imports', async () => {
-      let testOutput = ''
-      childProcess = exec(
-        runTestsCommand,
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '-r dd-trace/ci/init ' +
-              '--require ./ci-visibility/jest-mock-bypass-require/track-logger-resolution.js ' +
-              '--experimental-vm-modules',
-            TESTS_TO_RUN: 'jest-mock-bypass-require/esm-unrelated-cjs-test.mjs',
-            USE_JEST_RUN: '1',
-          },
-        }
-      )
-      childProcess.stdout.on('data', chunk => {
-        testOutput += chunk.toString()
-      })
-      childProcess.stderr.on('data', chunk => {
-        testOutput += chunk.toString()
-      })
-
-      const [code] = await once(childProcess, 'exit')
-      assert.doesNotMatch(testOutput, /\[unexpected logger resolution\]/)
-      assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-    })
-
-    esmIt('instruments a statically imported logger through a renamed symlink', async () => {
-      const loggerModulePath = path.join(cwd, 'node_modules', 'winston')
-      const linkedLoggerPath = path.join(cwd, 'linked-logger')
-      const linkedLoggerModulePath = path.join(linkedLoggerPath, 'node_modules', 'winston')
-      const linkedLoggerIndexPath = path.join(linkedLoggerPath, 'index.js')
-      const linkedLoggerPackagePath = path.join(linkedLoggerPath, 'package.json')
-      fs.mkdirSync(path.dirname(linkedLoggerModulePath), { recursive: true })
-      fs.renameSync(loggerModulePath, linkedLoggerModulePath)
-      fs.writeFileSync(linkedLoggerIndexPath, "module.exports = require('./node_modules/winston')\n")
-      fs.writeFileSync(linkedLoggerPackagePath, '{"name":"winston","main":"index.js"}\n')
-
-      try {
-        fs.symlinkSync(linkedLoggerPath, loggerModulePath, 'junction')
-        assert.strictEqual(fs.realpathSync(loggerModulePath), fs.realpathSync(linkedLoggerPath))
-
-        let testOutput = ''
-        const logsPromise = receiver
-          .gatherPayloadsMaxTimeout(({ url }) => url.includes('/api/v2/logs'), payloads => {
-            assert.strictEqual(payloads.length, 1, testOutput)
-
-            const [{ headers, logMessage, url }] = payloads
-            assert.strictEqual(headers['content-type'], 'application/json')
-            assert.strictEqual(headers['dd-api-key'], 'api-key')
-            assert.strictEqual(url, '/api/v2/logs?ddsource=winston&service=my-service')
-            assert.strictEqual(logMessage.length, 1)
-
-            const [{ dd, message }] = logMessage
-            assert.strictEqual(message, 'linked logger')
-            assert.strictEqual(dd.service, 'my-service')
-            assert.match(dd.trace_id, /^\d+$/)
-            assert.match(dd.span_id, /^\d+$/)
-          })
-
-        childProcess = exec(
-          runTestsCommand,
-          {
-            cwd,
-            env: {
-              ...getCiVisAgentlessConfig(receiver.port),
-              DD_AGENTLESS_LOG_SUBMISSION_ENABLED: '1',
-              DD_AGENTLESS_LOG_SUBMISSION_URL: `http://localhost:${receiver.port}`,
-              DD_API_KEY: 'api-key',
-              DD_SERVICE: 'my-service',
-              NODE_OPTIONS: '-r dd-trace/ci/init --experimental-vm-modules',
-              TESTS_TO_RUN: 'jest-mock-bypass-require/esm-linked-logger-test.mjs',
-              USE_JEST_RUN: '1',
-            },
-          }
-        )
-        childProcess.stdout.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-        childProcess.stderr.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-
-        const [[code]] = await Promise.all([
-          once(childProcess, 'exit'),
-          logsPromise,
-        ])
-
-        assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-      } finally {
-        if (fs.existsSync(loggerModulePath)) fs.unlinkSync(loggerModulePath)
-        fs.renameSync(linkedLoggerModulePath, loggerModulePath)
-        fs.unlinkSync(linkedLoggerIndexPath)
-        fs.unlinkSync(linkedLoggerPackagePath)
-        fs.rmdirSync(path.join(linkedLoggerPath, 'node_modules'))
-        fs.rmdirSync(linkedLoggerPath)
-      }
-    })
-
-    for (const loggerName of ['winston', 'pino', 'bunyan']) {
-      esmIt(`respects Jest ESM mocks for ${loggerName}`, async () => {
-        let testOutput = ''
-        childProcess = exec(
-          runTestsCommand,
-          {
-            cwd,
-            env: {
-              ...getCiVisAgentlessConfig(receiver.port),
-              NODE_OPTIONS: '-r dd-trace/ci/init --experimental-vm-modules',
-              TEST_LOGGER: loggerName,
-              TESTS_TO_RUN: 'jest-mock-bypass-require/esm-mock-test.mjs',
-              USE_JEST_RUN: '1',
-            },
-          }
-        )
-        childProcess.stdout.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-        childProcess.stderr.on('data', chunk => {
-          testOutput += chunk.toString()
-        })
-
-        const [code] = await once(childProcess, 'exit')
-        assert.strictEqual(code, 0, `Jest should pass but failed with code ${code}: ${testOutput}`)
-      })
-    }
   })
 
   context('seed suffix normalization', () => {
